// prodis-prodiswcl/src/app/modules/prodisapi/model/contratti.ts
/*
* SPDX-FileCopyrightText: Copyright 2022 | Regione Piemonte
* SPDX-License-Identifier: EUPL-1.2
*/
/**
 * Prodis
 * API per il backend della suite di Prodis.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { LavoratoriInForza } from './lavoratoriInForza';
import { LavoratoriSilp } from './lavoratoriSilp';


export interface Contratti { 
    codTipoContrattoMin?: string;
    dsContratto?: string;
    dtFine?: Date;
    dtInizio?: Date;
    dtTmst?: Date;
    flgForma?: string;
    id?: number;
    lavoratoriInForzas?: Array<LavoratoriInForza>;
    lavoratoriSilps?: Array<LavoratoriSilp>;
    status?: string;
    tipo?: string;
    tipologiaRapporto?: string;
}