// prodis-prodiswcl/src/app/modules/prodisapi/model/istat2001livello5.ts
/*
* SPDX-FileCopyrightText: Copyright 2022 | Regione Piemonte
* SPDX-License-Identifier: EUPL-1.2
*/
/**
 * Prodis
 * API per il backend della suite di Prodis.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { LavoratoriInForza } from './lavoratoriInForza';
import { LavoratoriSilp } from './lavoratoriSilp';
import { PostiLavoroDisp } from './postiLavoroDisp';


export interface Istat2001livello5 { 
    codIstat2001livello5Min?: string;
    dsComIstat2001livello5?: string;
    dtFine?: Date;
    dtInizio?: Date;
    dtTmst?: Date;
    flgVldUd?: string;
    id?: number;
    idNewIstat?: number;
    lavoratoriInForzas?: Array<LavoratoriInForza>;
    lavoratoriSilps?: Array<LavoratoriSilp>;
    postiLavoroDisps?: Array<PostiLavoroDisp>;
}